requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
distro==1.8.0
Django==4.2.7
django-cors-headers==4.3.1
django-google-maps==0.13.0
django-phonenumber-field==7.2.0
django-rest-framework==0.1.0
django-twilio==0.14.3.2
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
inflection==0.5.1
multidict==6.0.4
openai==0.28.0
packaging==23.2
phonenumbers==8.13.25
Pillow==10.1.0
pydantic==2.5.1
pydantic_core==2.14.3
PyJWT==2.8.0
pyotp==2.9.0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
twilio==7.17.0
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
yarl==1.9.2
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
cachetools==5.3.3
certifi==2023.7.22
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.1.2
debugpy==1.6.6
decorator==5.1.1
distro==1.8.0
Django==4.2.7
django-cors-headers==4.3.1
django-google-maps==0.13.0
django-phonenumber-field==7.2.0
django-rest-framework==0.1.0
django-twilio==0.14.3.2
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
executing==1.2.0
frozenlist==1.4.0
google-auth==2.31.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
inflection==0.5.1
ipykernel==6.21.2
ipython==8.10.0
jedi==0.18.2
jupyter_client==8.0.2
jupyter_core==5.2.0
matplotlib-inline==0.1.6
msgpack==1.0.8
multidict==6.0.4
nest-asyncio==1.5.6
openai==0.28.0
packaging==23.2
parso==0.8.3
phonenumbers==8.13.25
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.0.0
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.5.1
pydantic_core==2.14.3
pyfcm==2.0.4
Pygments==2.14.0
PyJWT==2.8.0
pyotp==2.9.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
pytz==2023.3.post1
pywin32==305
PyYAML==6.0.1
pyzmq==25.0.0
redis==5.0.7
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.2
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
twilio==7.17.0
typing_extensions==4.8.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.1.0
wcwidth==0.2.6
yarl==1.9.2
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
cachetools==5.3.3
certifi==2023.7.22
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.1.2
debugpy==1.6.6
decorator==5.1.1
distro==1.8.0
Django==4.2.7
django-cors-headers==4.3.1
django-google-maps==0.13.0
django-phonenumber-field==7.2.0
django-rest-framework==0.1.0
django-twilio==0.14.3.2
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
executing==1.2.0
frozenlist==1.4.0
google-auth==2.31.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
inflection==0.5.1
ipykernel==6.21.2
ipython==8.10.0
jedi==0.18.2
jupyter_client==8.0.2
jupyter_core==5.2.0
matplotlib-inline==0.1.6
msgpack==1.0.8
multidict==6.0.4
nest-asyncio==1.5.6
openai==0.28.0
packaging==23.2
parso==0.8.3
phonenumbers==8.13.25
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.0.0
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.5.1
pydantic_core==2.14.3
pyfcm==2.0.4
Pygments==2.14.0
PyJWT==2.8.0
pyotp==2.9.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
pytz==2023.3.post1
pywin32==305
PyYAML==6.0.1
pyzmq==25.0.0
redis==5.0.7
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.2
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
twilio==7.17.0
typing_extensions==4.8.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.1.0
wcwidth==0.2.6
yarl==1.9.2
